New weapon - Boomer - and effect - LongExplosion. Focuses more on effect damage than impact damage. Want to make this a trigger release later (remote bomb).

+++ GameApp/src/gameapp/Projectile/Spread.java
@@ -8,7 +8,7 @@ import javax.swing.ImageIcon;
  */
 public class Spread extends Projectile{
     
-    public Spread(int x, int y, int run, int rise, boolean up){
+    public Spread(int x, int y, double run, double rise, boolean up){
         super(x, y, up);
         ImageIcon icon = new ImageIcon("images/BurstShot.gif");
